// src/types/types.ts
type QueryParams = {
    filter?: Record<string, any>;
    search?: string;
    searchField?: string;
    sort?: string;
    select?: string[];
    sample?: { count: number; seed?: number };
    groupBy?: string;
    page?: number;
    perPage?: number;
    cursor?: string;
}

type Meta = {
    count: number;
    db_response_time_ms: number;
    page?: number;
    per_page?: number;
    next_cursor?: string;
}